+++ org.eclipse.jdt.core/compiler/org/eclipse/jdt/internal/compiler/impl/LongConstant.java
@@ -22,16 +22,16 @@ public char charValue() {
 	return (char) value;
 }
 public double doubleValue() {
-	return (double) value;
+	return value; // implicit cast to return type
 }
 public float floatValue() {
-	return (float) value;
+	return value; // implicit cast to return type
 }
 public int intValue() {
 	return (int) value;
 }
 public long longValue() {
-	return (long) value;
+	return value; 
 }
 public short shortValue() {
 	return (short) value;
