+++ org.eclipse.jdt.core/compiler/org/eclipse/jdt/internal/compiler/impl/ShortConstant.java
@@ -22,19 +22,19 @@ public char charValue() {
 	return (char) value;
 }
 public double doubleValue() {
-	return (double) value;
+	return value; // implicit cast to return type
 }
 public float floatValue() {
-	return (float) value;
+	return value; // implicit cast to return type
 }
 public int intValue() {
-	return (int) value;
+	return value; // implicit cast to return type
 }
 public long longValue() {
-	return (long) value;
+	return value; // implicit cast to return type
 }
 public short shortValue() {
-	return (short) value;
+	return value;
 }
 public String stringValue() {
 	//spec 15.17.11
